Move vulnerability analysis logic into vulnerability_summarizer.py and add tests of version comparison logic.
