+++ tests/test_vulnerability_summarizer.py
@@ -1,0 +1,33 @@
+import os
+import re
+import sys
+import json
+
+from addict import Dict
+
+import vulnerability_summarizer
+
+
+VERSION_TESTS = [
+    {"original": "= 1.2.3", "uniform": "1.2.3", "sortable": "0001.0002.0003"},
+    {"original": "= 1.12.3", "uniform": "1.12.3", "sortable": "0001.0012.0003"},
+    {"original": "= 1.12.20", "uniform": "1.12.20", "sortable": "0001.0012.0020"},
+]
+
+
+def test_get_uniform_version():
+    for version in VERSION_TESTS:
+        assert version["uniform"] == vulnerability_summarizer.get_uniform_version(
+            version["original"]
+        )
+
+
+def test_get_sortable_version():
+    for version in VERSION_TESTS:
+        assert version["sortable"] == vulnerability_summarizer.get_sortable_version(
+            version["original"]
+        )
+
+    assert vulnerability_summarizer.get_sortable_version(
+        "1.2.3"
+    ) < vulnerability_summarizer.get_sortable_version("1.12.3")
